+++ src/main/java/sortdir/quicksorts/BusQuickSort.java
@@ -1,13 +1,11 @@
 package sortdir.quicksorts;
 
 import classes.Bus;
-import sortdir.QuickSort;
-import sortdir.QuickSortStrategy;
 import sortdir.comparators.BusComparator;
 
 import java.util.Comparator;
 
-public class QuickSortBus implements QuickSortStrategy<Bus> {
+public class BusQuickSort implements QuickSortStrategy<Bus> {
     @Override
     public Bus[] sort(Bus[] arr) {
