+++ src/main/java/sortdir/quicksorts/QuickSort.java
@@ -1,4 +1,4 @@
-package sortdir;
+package sortdir.quicksorts;
 
 import java.util.Arrays;
 import java.util.Comparator;
