+++ src/main/java/sortdir/quicksorts/QuickSortStrategy.java
@@ -1,4 +1,4 @@
-package sortdir;
+package sortdir.quicksorts;
 
 public interface QuickSortStrategy<T> {
     T[] sort(T[] arr);
