+++ src/main/java/sortdir/quicksorts/StudentQuickSort.java
@@ -1,13 +1,11 @@
 package sortdir.quicksorts;
 
 import classes.Student;
-import sortdir.QuickSort;
-import sortdir.QuickSortStrategy;
 import sortdir.comparators.StudentComparator;
 
 import java.util.Comparator;
 
-public class QuickSortStudent implements QuickSortStrategy<Student> {
+public class StudentQuickSort implements QuickSortStrategy<Student> {
     @Override
     public Student[] sort(Student[] arr) {
